# centinela-agents/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Centinela · API de agentes antifraude",
    "description": "API de agentes de **Centinela**, plataforma antifraude multimodal.\n\nDos procesos, un mismo contrato de salida (`Decision`):\n\n- `POST /api/v1/transactions/evaluate` — aprobar · validación adicional · bloquear\n- `POST /api/v1/documents/validate` — auténtico · sospechoso · falso\n\nToda decisión es explicable y trazable bajo un `trace_id`. Los casos bloqueados\ny las excepciones siempre pasan por una persona.\n\n**Datos sintéticos.** Este servicio se construyó como proyecto académico y no\ndebe procesar datos reales de clientes sin gobierno y consentimiento.\n",
    "contact": {
      "name": "Grupo Centinela"
    },
    "license": {
      "name": "MIT"
    },
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/v1/documents/validate": {
      "post": {
        "tags": [
          "documentos"
        ],
        "summary": "Valida un documento y determina si es auténtico, sospechoso o falso",
        "operationId": "validate_api_v1_documents_validate_post",
        "requestBody": {
          "content": {
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/Body_validate_api_v1_documents_validate_post"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DocumentDecision"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/documents/references": {
      "post": {
        "tags": [
          "documentos"
        ],
        "summary": "Registra una firma o plantilla de referencia",
        "operationId": "register_reference_api_v1_documents_references_post",
        "requestBody": {
          "content": {
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/Body_register_reference_api_v1_documents_references_post"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ReferenceResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/documents/{trace_id}/evidence": {
      "get": {
        "tags": [
          "documentos"
        ],
        "summary": "Regiones de los hallazgos para dibujar sobre la imagen",
        "operationId": "evidence_api_v1_documents__trace_id__evidence_get",
        "parameters": [
          {
            "name": "trace_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Trace Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EvidenceRegionsResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/documents/{trace_id}/page/{page}": {
      "get": {
        "tags": [
          "documentos"
        ],
        "summary": "Imagen procesada de una página (solo analista/supervisor)",
        "operationId": "page_image_api_v1_documents__trace_id__page__page__get",
        "parameters": [
          {
            "name": "trace_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Trace Id"
            }
          },
          {
            "name": "page",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Page"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Página renderizada",
            "content": {
              "image/png": {}
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/cases": {
      "get": {
        "tags": [
          "casos"
        ],
        "summary": "Bandeja de casos",
        "operationId": "list_cases_api_v1_cases_get",
        "parameters": [
          {
            "name": "process",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "enum": [
                    "transaction",
                    "document"
                  ],
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Process"
            }
          },
          {
            "name": "verdict",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Verdict"
            }
          },
          {
            "name": "requires_human_review",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "boolean"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Requires Human Review"
            }
          },
          {
            "name": "date_from",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "format": "date-time"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Date From"
            }
          },
          {
            "name": "date_to",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "format": "date-time"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Date To"
            }
          },
          {
            "name": "order_by",
            "in": "query",
            "required": false,
            "schema": {
              "enum": [
                "score",
                "created_at"
              ],
              "type": "string",
              "default": "score",
              "title": "Order By"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 500,
              "minimum": 1,
              "default": 100,
              "title": "Limit"
            }
          },
          {
            "name": "offset",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 0,
              "default": 0,
              "title": "Offset"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/CaseSummary"
                  },
                  "title": "Response List Cases Api V1 Cases Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/cases/{trace_id}": {
      "get": {
        "tags": [
          "casos"
        ],
        "summary": "Detalle completo de un caso, con la traza",
        "operationId": "get_case_api_v1_cases__trace_id__get",
        "parameters": [
          {
            "name": "trace_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Trace Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CaseDetail"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/feedback": {
      "post": {
        "tags": [
          "casos"
        ],
        "summary": "Etiqueta del analista; indexa el caso para que el agente aprenda",
        "operationId": "post_feedback_api_v1_feedback_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Feedback"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FeedbackAck"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "Body_register_reference_api_v1_documents_references_post": {
        "properties": {
          "file": {
            "type": "string",
            "contentMediaType": "application/octet-stream",
            "title": "File"
          },
          "document_type": {
            "type": "string",
            "title": "Document Type",
            "default": "firma"
          },
          "label": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Label",
            "description": "Etiqueta libre, sin datos personales"
          }
        },
        "type": "object",
        "required": [
          "file"
        ],
        "title": "Body_register_reference_api_v1_documents_references_post"
      },
      "Body_validate_api_v1_documents_validate_post": {
        "properties": {
          "file": {
            "type": "string",
            "contentMediaType": "application/octet-stream",
            "title": "File",
            "description": "JPG, PNG o PDF de hasta 10 MB."
          },
          "document_type": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Document Type"
          },
          "reference_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Reference Id"
          },
          "expected_fields": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Expected Fields",
            "description": "JSON campo → valor"
          }
        },
        "type": "object",
        "required": [
          "file"
        ],
        "title": "Body_validate_api_v1_documents_validate_post"
      },
...
      "CaseSummary": {
        "properties": {
          "trace_id": {
            "type": "string",
            "title": "Trace Id"
          },
          "process": {
            "type": "string",
            "enum": [
              "transaction",
              "document"
            ],
            "title": "Process"
          },
          "verdict": {
            "type": "string",
            "enum": [
              "aprobar",
              "validacion_adicional",
              "bloquear",
              "autentico",
              "sospechoso",
              "falso"
            ],
            "title": "Verdict"
          },
          "score": {
            "type": "number",
            "title": "Score"
          },
          "confidence": {
            "type": "number",
            "title": "Confidence"
          },
          "requires_human_review": {
            "type": "boolean",
            "title": "Requires Human Review"
          },
          "shadow": {
            "type": "boolean",
            "title": "Shadow",
            "default": false
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          },
          "cost_usd": {
            "type": "number",
            "title": "Cost Usd",
            "default": 0.0
          },
          "latency_ms": {
            "type": "integer",
            "title": "Latency Ms",
            "default": 0
          },
          "feedback_label": {
            "anyOf": [
              {
                "type": "string",
                "enum": [
                  "fraude_confirmado",
                  "legitimo",
                  "documento_falso",
                  "documento_autentico"
                ]
              },
              {
                "type": "null"
              }
            ],
            "title": "Feedback Label"
          },
          "age_seconds": {
            "type": "number",
            "title": "Age Seconds",
            "default": 0.0
          }
        },
        "type": "object",
        "required": [
          "trace_id",
          "process",
          "verdict",
          "score",
          "confidence",
          "requires_human_review",
          "created_at"
        ],
        "title": "CaseSummary",
        "description": "Fila de la bandeja del analista."
      },
...
      "DocumentDecision": {
        "properties": {
          "trace_id": {
            "type": "string",
            "title": "Trace Id"
          },
          "process": {
            "type": "string",
            "enum": [
              "transaction",
              "document"
            ],
            "title": "Process"
          },
          "verdict": {
            "type": "string",
            "enum": [
              "aprobar",
              "validacion_adicional",
              "bloquear",
              "autentico",
              "sospechoso",
              "falso"
            ],
            "title": "Verdict"
          },
          "score": {
            "type": "number",
            "maximum": 1.0,
            "minimum": 0.0,
            "title": "Score"
          },
          "confidence": {
            "type": "number",
            "maximum": 1.0,
            "minimum": 0.0,
            "title": "Confidence"
          },
          "reasons": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Reasons"
          },
          "evidence": {
            "items": {
              "$ref": "#/components/schemas/Evidence"
            },
            "type": "array",
            "title": "Evidence"
          },
          "requires_human_review": {
            "type": "boolean",
            "title": "Requires Human Review",
            "default": false
          },
          "shadow": {
            "type": "boolean",
            "title": "Shadow",
            "description": "true si SHADOW_MODE forzó el veredicto a `aprobar`.",
            "default": false
          },
          "explanation_customer": {
            "type": "string",
            "title": "Explanation Customer",
            "description": "Texto breve y amable. Jamás menciona reglas, umbrales, modelos, scores ni hallazgos forenses.",
            "default": ""
          },
          "explanation_analyst": {
            "type": "string",
            "title": "Explanation Analyst",
            "description": "Texto técnico para el analista.",
            "default": ""
          },
          "model": {
            "type": "string",
            "title": "Model",
            "default": ""
          },
          "tokens_in": {
            "type": "integer",
            "title": "Tokens In",
            "default": 0
          },
          "tokens_out": {
            "type": "integer",
            "title": "Tokens Out",
            "default": 0
          },
          "cost_usd": {
            "type": "number",
            "title": "Cost Usd",
            "default": 0.0
          },
          "latency_ms": {
            "type": "integer",
            "title": "Latency Ms",
            "default": 0
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          },
          "document_type_detected": {
            "type": "string",
            "enum": [
              "cedula",
              "comprobante_domicilio",
              "contrato",
              "poder",
              "liquidacion",
              "firma",
              "otro"
            ],
            "title": "Document Type Detected",
            "default": "otro"
          },
          "fields": {
            "additionalProperties": {
              "$ref": "#/components/schemas/FieldValue"
            },
            "type": "object",
            "title": "Fields"
          },
          "checks": {
            "items": {
              "$ref": "#/components/schemas/Check"
            },
            "type": "array",
            "title": "Checks"
          },
          "findings": {
            "items": {
              "$ref": "#/components/schemas/Finding"
            },
            "type": "array",
            "title": "Findings"
          },
          "pages_analyzed": {
            "type": "integer",
            "title": "Pages Analyzed",
            "default": 1
          },
          "image_quality": {
            "type": "number",
            "maximum": 1.0,
            "minimum": 0.0,
            "title": "Image Quality",
            "default": 1.0
          }
        },
        "type": "object",
        "required": [
          "trace_id",
          "process",
          "verdict",
          "score",
          "confidence"
        ],
        "title": "DocumentDecision",
        "description": "`Decision` + lo específico del proceso B.\n\nVive en este módulo, junto a `Decision` y `CaseDetail`, porque los tres se\nreferencian entre sí: `CaseDetail.decision` debe poder ser un\n`DocumentDecision` completo para que la pantalla del analista reciba\n`checks`, `fields` y `findings`."
      },
...
      "EvidenceRegionsResponse": {
        "properties": {
          "trace_id": {
            "type": "string",
            "title": "Trace Id"
          },
          "pages_analyzed": {
            "type": "integer",
            "title": "Pages Analyzed"
          },
          "image_quality": {
            "type": "number",
            "title": "Image Quality"
          },
          "findings": {
            "items": {
              "$ref": "#/components/schemas/Finding"
            },
            "type": "array",
            "title": "Findings"
          },
          "artifacts_available": {
            "type": "boolean",
            "title": "Artifacts Available",
            "description": "false cuando la retención ya borró las imágenes procesadas.",
            "default": false
          }
        },
        "type": "object",
        "required": [
          "trace_id",
          "pages_analyzed",
          "image_quality"
        ],
        "title": "EvidenceRegionsResponse",
        "description": "Regiones que el front dibuja como recuadros sobre la imagen."
      },
      "Feedback": {
        "properties": {
          "trace_id": {
            "type": "string",
            "title": "Trace Id"
          },
          "label": {
            "type": "string",
            "enum": [
              "fraude_confirmado",
              "legitimo",
              "documento_falso",
              "documento_autentico"
            ],
            "title": "Label"
          },
          "analyst": {
            "type": "string",
            "minLength": 1,
            "title": "Analyst",
            "description": "Identificador del analista."
          },
          "comment": {
            "type": "string",
            "title": "Comment",
            "default": ""
          }
        },
        "type": "object",
        "required": [
          "trace_id",
          "label",
          "analyst"
        ],
        "title": "Feedback",
        "description": "Etiqueta del analista. Cierra el ciclo de aprendizaje: indexa el caso."
      },
      "FeedbackAck": {
        "properties": {
          "trace_id": {
            "type": "string",
            "title": "Trace Id"
          },
          "label": {
            "type": "string",
            "enum": [
              "fraude_confirmado",
              "legitimo",
              "documento_falso",
              "documento_autentico"
            ],
            "title": "Label"
          },
          "indexed": {
            "type": "boolean",
            "title": "Indexed",
            "description": "true si el caso se indexó en el vector store."
          },
          "message": {
            "type": "string",
            "title": "Message",
            "default": ""
          }
        },
        "type": "object",
        "required": [
          "trace_id",
          "label",
          "indexed"
        ],
        "title": "FeedbackAck"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "ReferenceResponse": {
        "properties": {
          "reference_id": {
            "type": "string",
            "title": "Reference Id"
          },
          "document_type": {
            "type": "string",
            "enum": [
              "cedula",
              "comprobante_domicilio",
              "contrato",
              "poder",
              "liquidacion",
              "firma",
              "otro"
            ],
            "title": "Document Type"
          },
          "stored_at": {
            "type": "string",
            "title": "Stored At"
          },
          "message": {
            "type": "string",
            "title": "Message",
            "default": ""
          }
        },
        "type": "object",
        "required": [
          "reference_id",
          "document_type",
          "stored_at"
        ],
        "title": "ReferenceResponse",
        "description": "Respuesta al registrar una firma o plantilla de referencia."
      },
...
    }
  }
}